requirements.txt
yt-dlp
ffmpeg-python
pandas
streamlit
sentence-transformers
faiss-cpu
pandas
numpy